[update] Move some auxilary function in test_on_sim_real.py to util script
